ts and bugs fixed
